Provide a usleep
This is possibly something that should be provided by a separate library,
and not in the BSP, but including it here will surely be helpful for users
just doing basic demos (as in our platform example). If it causes issues
clashing with a standard library, it can just be renamed.